tests: Move logs test cases into their own module
