+++ tests/utils/tests.py
@@ -2,13 +2,10 @@ import datetime
 import decimal
 import functools
 import importlib
-import io
 import json
-import logging
 import uuid
 from datetime import date
 from unittest import mock
-from unittest.mock import patch
 
 import faker
 import pytest
@@ -20,7 +17,6 @@ from django.contrib.auth.models import AnonymousUser, Permission
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.messages.middleware import MessageMiddleware
 from django.contrib.sessions.middleware import SessionMiddleware
-from django.core import management
 from django.core.exceptions import ValidationError
 from django.http import HttpResponse
 from django.template import Context, Template
@@ -1709,75 +1705,6 @@ def test_previous_step():
 
     res = render_to_string("layout/previous_step.html", {"back_url": "/search/results?back_url=/blabla&other=params"})
     assert PREVIOUS_IS_LIST in res
-
-
-def test_log_current_organization(client):
-    membership = CompanyMembershipFactory()
-    client.force_login(membership.user)
-    root_logger = logging.getLogger()
-    stream_handler = root_logger.handlers[0]
-    captured = io.StringIO()
-    assert isinstance(stream_handler, logging.StreamHandler)
-    # caplog cannot be used since the organization_id is written by the log formatter
-    # capsys/capfd did not want to work because https://github.com/pytest-dev/pytest/issues/5997
-    with patch.object(stream_handler, "stream", captured):
-        response = client.get(reverse("dashboard:index"))
-    assert response.status_code == 200
-    # Check that the organization_id is properly logged to stdout
-    assert f'"usr.organization_id": {membership.company_id}' in captured.getvalue()
-
-
-def test_log_current_command(client):
-    root_logger = logging.getLogger()
-    stream_handler = root_logger.handlers[0]
-    captured = io.StringIO()
-    assert isinstance(stream_handler, logging.StreamHandler)
-    # caplog cannot be used since the command infos are written by the log formatter
-    # capsys/capfd did not want to work because https://github.com/pytest-dev/pytest/issues/5997
-    with patch.object(stream_handler, "stream", captured):
-        management.call_command(
-            # This could have been any other command inheriting from LoggedCommandMixin
-            "delete_old_emails"
-        )
-    # Check that the organization_id is properly logged to stdout
-    lines = captured.getvalue().splitlines()
-    log = json.loads(lines[0])
-    assert log["command.name"] == "itou.emails.management.commands.delete_old_emails"
-    assert "command.run_uid" in log
-    assert uuid.UUID(log["command.run_uid"])
-
-
-def test_log_hijack_infos(client):
-    LOG_KEY = "usr.hijack_history"
-    dashboard_url = reverse("dashboard:index")
-    membership = CompanyMembershipFactory()
-    client.force_login(membership.user)
-    root_logger = logging.getLogger()
-    stream_handler = root_logger.handlers[0]
-    captured = io.StringIO()
-    assert isinstance(stream_handler, logging.StreamHandler)
-    # caplog cannot be used since the organization_id is written by the log formatter
-    # capsys/capfd did not want to work because https://github.com/pytest-dev/pytest/issues/5997
-    with patch.object(stream_handler, "stream", captured):
-        response = client.get(dashboard_url)
-    assert response.status_code == 200
-    # Check that the hijack info is not there
-    assert f'"usr.id": {membership.user.id}' in captured.getvalue()
-    assert LOG_KEY not in captured.getvalue()
-
-    hijacker = ItouStaffFactory(is_superuser=True)
-    client.force_login(hijacker)
-    response = client.post(reverse("hijack:acquire"), {"user_pk": membership.user.pk, "next": dashboard_url})
-    assertRedirects(response, dashboard_url, fetch_redirect_response=False)
-    captured = io.StringIO()
-    # caplog cannot be used since the organization_id is written by the log formatter
-    # capsys/capfd did not want to work because https://github.com/pytest-dev/pytest/issues/5997
-    with patch.object(stream_handler, "stream", captured):
-        response = client.get(dashboard_url)
-    assert response.status_code == 200
-    # Check that the hijack info is there
-    assert f'"usr.id": {membership.user.id}' in captured.getvalue()
-    assert f'"{LOG_KEY}": ["{hijacker.pk}"]' in captured.getvalue()
 
 
 def test_create_fake_postcode():
